chore: rename datatypes to data_types
