+++ autonomi/src/client/data_types/scratchpad.rs
